+++ polynote-frontend/polynote/state/server_state.ts
@@ -52,6 +52,7 @@ export interface ServerState {
     // Keys are notebook path. Values denote whether the notebook has ever been loaded in this session.
     notebooks: Record<string, NotebookInfo["loaded"]>,
     dependencySources: Record<string, DependencySource>,
+    notebookTimestamps: Record<string, number>,
     connectionStatus: "connected" | "disconnected",
     interpreters: Record<string, string>,
     serverVersion: string,
@@ -80,6 +81,7 @@ export class ServerStateHandler extends BaseHandler<ServerState> {
             ServerStateHandler.inst = new ServerStateHandler(new ObjectStateHandler<ServerState>({
                 notebooks: {},
                 dependencySources: {},
+                notebookTimestamps: {},
                 connectionStatus: "disconnected",
                 interpreters: {},
                 serverVersion: "unknown",
@@ -126,6 +128,7 @@ export class ServerStateHandler extends BaseHandler<ServerState> {
             ServerStateHandler.inst = new ServerStateHandler(new ObjectStateHandler<ServerState>({
                 notebooks: {},
                 dependencySources: {},
+                notebookTimestamps: {},
                 connectionStatus: "disconnected",
                 interpreters: {},
                 serverVersion: "unknown",
@@ -210,7 +213,8 @@ export class ServerStateHandler extends BaseHandler<ServerState> {
                 const pathIdx = state.openFiles.findIndex(of => of.type === 'notebook' && of.path === oldPath);
                 return {
                     notebooks: renameKey(oldPath, newPath),
-                    openFiles: pathIdx >= 0 ? replaceArrayValue({type: 'notebook', newPath}, pathIdx) : NoUpdate
+                    openFiles: pathIdx >= 0 ? replaceArrayValue({type: 'notebook', newPath}, pathIdx) : NoUpdate,
+                    notebookTimestamps: renameKey(oldPath, newPath),
                 }
             })
         }
@@ -218,8 +222,9 @@ export class ServerStateHandler extends BaseHandler<ServerState> {
 
     static deleteNotebook(path: string) {
         ServerStateHandler.closeFile(path, /*reinitialize*/ false).then(() => {
-            // update the server state's notebook dictionary
+            // update the server state's notebook dictionaries
             ServerStateHandler.get.updateField("notebooks", notebooks => notebooks[path] !== undefined ? removeKey(path) : NoUpdate);
+            ServerStateHandler.get.updateField("notebookTimestamps", notebooks => notebooks[path] !== undefined ? removeKey(path) : NoUpdate);
         })
     }