+++ polynote-server/src/main/scala/polynote/server/NotebookSession.scala
@@ -1,7 +1,7 @@
 package polynote.server
 
 import cats.syntax.either._
-import polynote.kernel.environment.{Env, PublishMessage}
+import polynote.kernel.environment.{BroadcastAll, BroadcastMessage, Env, PublishMessage}
 import polynote.kernel.logging.Logging
 import polynote.kernel.util.{Publish, UPublish}
 import polynote.kernel.{BaseEnv, Complete, GlobalEnv, PresenceUpdate, Running, StreamingHandles, TaskInfo, UpdatedTasks}
@@ -13,6 +13,7 @@ import uzhttp.websocket.Frame
 import zio.stream.{Stream, Take, UStream, ZStream}
 import zio.{Promise, RIO, UIO, ZIO, ZLayer, ZManaged, ZQueue}
 
+import java.util.Calendar
 import java.util.concurrent.atomic.AtomicInteger
 
 
@@ -176,11 +177,12 @@ class NotebookSession(
 
 object NotebookSession {
 
-  def stream(path: String, input: Stream[Throwable, Frame], broadcastAll: UPublish[Message]): ZManaged[SessionEnv with NotebookManager, HTTPError, UStream[Frame]] = {
+  def stream(path: String, input: Stream[Throwable, Frame], broadcastAll: BroadcastMessage): ZManaged[SessionEnv with NotebookManager, HTTPError, UStream[Frame]] = {
     for {
       _                <- NotebookManager.assertValidPath(path).toManaged_
       output           <- ZQueue.unbounded[Take[Nothing, Message]].toManaged_ // TODO: finalizer instead of close
       _                <- Env.addManaged[SessionEnv with NotebookManager](Publish(output))
+      _                <- Env.addManaged[SessionEnv with NotebookManager with PublishMessage](broadcastAll)
       subscriber       <- NotebookManager.subscribe(path).orElseFail(NotFound(path))
       sessionId        <- nextSessionId.toManaged_
       streamingHandles <- StreamingHandles.make(sessionId).orDie.toManaged_