+++ polynote-server/src/main/scala/polynote/server/package.scala
@@ -1,11 +1,11 @@
 package polynote
 
 import polynote.app.MainArgs
-import polynote.kernel.environment.{Config, PublishMessage}
+import polynote.kernel.environment.{BroadcastAll, Config, PublishMessage}
 import polynote.kernel.logging.Logging
 import polynote.kernel.util.{RefMap, UPublish}
 import polynote.kernel.{BaseEnv, GlobalEnv, KernelBusyState}
-import polynote.messages.{CreateNotebook, DeleteNotebook, Message, NotebookCell, NotebookSearchResult, RenameNotebook, ShortString}
+import polynote.messages.{CreateNotebook, DeleteNotebook, Message, NotebookCell, NotebookSearchResult, RenameNotebook, ShortString, fsNotebook}
 import polynote.server.auth.{IdentityProvider, UserIdentity}
 import polynote.server.repository.format.NotebookFormat
 import polynote.server.repository.fs.FileSystems
@@ -108,28 +108,28 @@ package object server {
     }
 
     def access: URIO[NotebookManager, Service] = ZIO.access[NotebookManager](_.get)
-    def open(path: String): RIO[NotebookManager with BaseEnv with GlobalEnv, KernelPublisher] = access.flatMap(_.open(path))
-    def subscribe(path: String): RManaged[NotebookManager with BaseEnv with GlobalEnv with PublishMessage with UserIdentity, KernelSubscriber] = open(path).toManaged_.flatMap(_.subscribe())
+    def open(path: String): RIO[NotebookManager with BaseEnv with GlobalEnv with BroadcastAll, KernelPublisher] = access.flatMap(_.open(path))
+    def subscribe(path: String): RManaged[NotebookManager with BaseEnv with GlobalEnv with PublishMessage with BroadcastAll with UserIdentity, KernelSubscriber] = open(path).toManaged_.flatMap(_.subscribe())
     def fetchIfOpen(path: String): RIO[NotebookManager with BaseEnv with GlobalEnv, Option[(String, String)]] = access.flatMap(_.fetchIfOpen(path))
     def location(path: String): RIO[NotebookManager with BaseEnv with GlobalEnv, Option[URI]] = access.flatMap(_.location(path))
-    def list(): RIO[NotebookManager with BaseEnv with GlobalEnv, List[String]] = access.flatMap(_.list())
+    def list(): RIO[NotebookManager with BaseEnv with GlobalEnv, List[fsNotebook]] = access.flatMap(_.list())
     def listRunning(): RIO[NotebookManager with BaseEnv with GlobalEnv, List[String]] = access.flatMap(_.listRunning())
     def status(path: String): RIO[NotebookManager with BaseEnv with GlobalEnv, KernelBusyState] = access.flatMap(_.status(path))
     def create(path: String, maybeContent: Option[String]): RIO[NotebookManager with BaseEnv with GlobalEnv, String] = access.flatMap(_.create(path, maybeContent))
-    def rename(path: String, newPath: String): RIO[NotebookManager with BaseEnv with GlobalEnv, String] = access.flatMap(_.rename(path, newPath))
+    def rename(path: String, newPath: String): RIO[NotebookManager with BaseEnv with GlobalEnv with BroadcastAll, String] = access.flatMap(_.rename(path, newPath))
     def copy(path: String, newPath: String): RIO[NotebookManager with BaseEnv with GlobalEnv, String] = access.flatMap(_.copy(path, newPath))
     def delete(path: String): RIO[NotebookManager with BaseEnv with GlobalEnv, Unit] = access.flatMap(_.delete(path))
     def search(query: String): RIO[NotebookManager with BaseEnv with GlobalEnv, List[NotebookSearchResult]] = access.flatMap(_.search(query))
 
     trait Service {
-      def open(path: String): RIO[BaseEnv with GlobalEnv, KernelPublisher]
+      def open(path: String): RIO[BaseEnv with GlobalEnv with BroadcastAll, KernelPublisher]
       def fetchIfOpen(path: String): RIO[BaseEnv with GlobalEnv, Option[(String, String)]]
       def location(path: String): RIO[BaseEnv with GlobalEnv, Option[URI]]
-      def list(): RIO[BaseEnv with GlobalEnv, List[String]]
+      def list(): RIO[BaseEnv with GlobalEnv, List[fsNotebook]]
       def listRunning(): RIO[BaseEnv with GlobalEnv, List[String]]
       def status(path: String): RIO[BaseEnv with GlobalEnv, KernelBusyState]
       def create(path: String, maybeContent: Option[String]): RIO[BaseEnv with GlobalEnv, String]
-      def rename(path: String, newPath: String): RIO[BaseEnv with GlobalEnv, String]
+      def rename(path: String, newPath: String): RIO[BaseEnv with GlobalEnv with BroadcastAll, String]
       def copy(path: String, newPath: String): RIO[BaseEnv with GlobalEnv, String]
       def delete(path: String): RIO[BaseEnv with GlobalEnv, Unit]
       def close(): RIO[BaseEnv, Unit]
@@ -160,7 +160,7 @@ package object server {
         private val maxRetryDelay = Duration(8, TimeUnit.SECONDS)
 
 
-        override def open(path: String): RIO[BaseEnv with GlobalEnv, KernelPublisher] = openNotebooks.getOrCreate(path) {
+        override def open(path: String): RIO[BaseEnv with GlobalEnv with BroadcastAll, KernelPublisher] = openNotebooks.getOrCreate(path) {
           for {
             notebookRef   <- repository.openNotebook(path)
             publisher     <- KernelPublisher(notebookRef, broadcastAll)
@@ -187,7 +187,7 @@ package object server {
 
         override def location(path: String): RIO[BaseEnv with GlobalEnv, Option[URI]] = repository.notebookURI(path)
 
-        override def list(): RIO[BaseEnv with GlobalEnv, List[String]] = repository.listNotebooks()
+        override def list(): RIO[BaseEnv with GlobalEnv, List[fsNotebook]] = repository.listNotebooks()
 
         override def listRunning(): RIO[BaseEnv, List[String]] = openNotebooks.keys
 
@@ -203,7 +203,7 @@ package object server {
             _        <- broadcastMessage(CreateNotebook(ShortString(realPath)))
           } yield realPath
 
-        override def rename(path: String, newPath: String): RIO[BaseEnv with GlobalEnv, String] =
+        override def rename(path: String, newPath: String): RIO[BaseEnv with GlobalEnv with BroadcastAll, String] =
           for {
             realPath <- openNotebooks.get(path).flatMap {
               case None            => repository.renameNotebook(path, newPath)
@@ -238,7 +238,7 @@ package object server {
         override def search(query: String): RIO[BaseEnv with GlobalEnv, List[NotebookSearchResult]] = {
           repository.listNotebooks.flatMap(nbs => ZIO.foreachParN(16)(nbs) { nb => {
             for {
-              loadedNB <- repository.loadNotebook(nb)
+              loadedNB <- repository.loadNotebook(nb.path)
               cells    <- ZIO(loadedNB.cells.filter(c => c.content.toString.contains(query)))
             } yield for {
               cell <- cells