+++ polynote-server/src/main/scala/polynote/server/repository/NotebookRepository.scala
@@ -6,7 +6,7 @@ import java.nio.file.{Path, Paths}
 import cats.implicits._
 import polynote.config.{Mount, PolynoteConfig, Wal}
 import polynote.kernel.NotebookRef.AlreadyClosed
-import polynote.kernel.environment.Config
+import polynote.kernel.environment.{BroadcastAll, Config}
 import polynote.kernel.{BaseEnv, GlobalEnv, NotebookRef, Result}
 import polynote.messages._
 import polynote.server.repository.fs.FileSystems
@@ -42,12 +42,12 @@ trait NotebookRepository {
   def saveNotebook(nb: Notebook): RIO[BaseEnv with GlobalEnv, Unit]
 
 
-  def openNotebook(path: String): RIO[BaseEnv with GlobalEnv, NotebookRef]
+  def openNotebook(path: String): RIO[BaseEnv with GlobalEnv with BroadcastAll, NotebookRef]
 
   /**
     * @return A list of notebook paths that exist in this repository
     */
-  def listNotebooks(): RIO[BaseEnv with GlobalEnv, List[String]]
+  def listNotebooks(): RIO[BaseEnv with GlobalEnv, List[fsNotebook]]
 
   /**
     * Create a notebook at the given path, optionally creating it from the given content string.
@@ -55,7 +55,7 @@ trait NotebookRepository {
     * TODO: Server no longer imports. Need to implement this on the client side.
     */
   def createNotebook(path: String, maybeContent: Option[String]): RIO[BaseEnv with GlobalEnv, String]
-  def createAndOpen(path: String, notebook: Notebook, version: Int = 0): RIO[BaseEnv with GlobalEnv, NotebookRef]
+  def createAndOpen(path: String, notebook: Notebook, version: Int = 0): RIO[BaseEnv with GlobalEnv with BroadcastAll, NotebookRef]
 
   def renameNotebook(path: String, newPath: String): RIO[BaseEnv with GlobalEnv, String]
 
@@ -158,7 +158,7 @@ class TreeRepository (
     override def clearAllResults(): IO[AlreadyClosed, List[CellID]] =
       withPermit(withRef(_.clearAllResults()))
 
-    override def rename(newPath: String): RIO[BaseEnv with GlobalEnv, String] =
+    override def rename(newPath: String): RIO[BaseEnv with GlobalEnv with BroadcastAll, String] =
       // The semaphore has Short.MaxValue permits so that the other operations can just take one and avoid blocking
       // each other. This operation takes all of them, to block concurrent renames and any other operations during rename.
       failIfClosed *> renameLock.withPermits(Short.MaxValue) {
@@ -229,7 +229,7 @@ class TreeRepository (
     * @param f:            Function for `(NotebookRepository, relativePath: String, maybeBasePath: Option[String) => RIO[Env, T]`.
     *                      The presence of `maybeBasePath` reflects whether the path has been relativized (in case callers need to de-relativize any results)
     */
-  private[repository] def delegate[T](notebookPath: String)(f: (NotebookRepository, String, Option[String]) => RIO[BaseEnv with GlobalEnv, T]): RIO[BaseEnv with GlobalEnv, T] = {
+  private[repository] def delegate[R, T](notebookPath: String)(f: (NotebookRepository, String, Option[String]) => RIO[R, T]): RIO[R, T] = {
     val (originalPath, basePath) = extractPath(Paths.get(reslash(notebookPath)))
 
     val (repoForPath, relativePath, maybeBasePath) = basePath.flatMap(base => repos.get(base.toString).map(_ -> base)) match {
@@ -277,7 +277,7 @@ class TreeRepository (
      } yield normalizePath(base.map(b => nb.copy(path = Paths.get(b, nb.path).toString)).getOrElse(nb))
   }
 
-  override def openNotebook(path: String): RIO[BaseEnv with GlobalEnv, NotebookRef] = delegate(path) {
+  override def openNotebook(path: String): RIO[BaseEnv with GlobalEnv with BroadcastAll, NotebookRef] = delegate(path) {
     (repo, relativePath, base) => repo.openNotebook(relativePath).flatMap {
       underlying => TreeNotebookRef(repo, underlying, base, relativePath)
     }
@@ -287,13 +287,17 @@ class TreeRepository (
     (repo, relativePath, _) => repo.saveNotebook(nb.copy(path = relativePath))
   }
 
-  override def listNotebooks(): RIO[BaseEnv with GlobalEnv, List[String]] = {
+  override def listNotebooks(): RIO[BaseEnv with GlobalEnv, List[fsNotebook]] = {
     for {
-      rootNBs  <- root.listNotebooks().map(_.map(deslash))
+      rootNBs  <- root.listNotebooks().map(_.map(nb => fsNotebook(deslash(nb.path), nb.lastSaved)))
       mountNbs <- ZIO.foreach(repos.toList) {
-        case (base, repo) => repo.listNotebooks().map(_.map(nbPath => deslash(Paths.get(base, nbPath).toString)))
+        case (base, repo) => repo.listNotebooks().map(_.map(nb => fsNotebook(deslash(Paths.get(base, nb.path).toString), nb.lastSaved)))
       }
-    } yield rootNBs ++ mountNbs.flatten
+      // flatten the nested list of mount notebooks
+    } yield rootNBs ++ (for {
+      nbs <- mountNbs
+      nb <- nbs
+    } yield nb)
   }
 
   override def createNotebook(originalPath: String, maybeContent: Option[String]): RIO[BaseEnv with GlobalEnv, String] = delegate(originalPath) {
@@ -303,7 +307,7 @@ class TreeRepository (
         } yield deslash(base.map(b => Paths.get(b, nbPath).toString).getOrElse(nbPath))
   }
 
-  override def createAndOpen(path: String, notebook: Notebook, version: Int): RIO[BaseEnv with GlobalEnv, NotebookRef] = delegate(path) {
+  override def createAndOpen(path: String, notebook: Notebook, version: Int): RIO[BaseEnv with GlobalEnv with BroadcastAll, NotebookRef] = delegate(path) {
     (repo, relativePath, base) =>
       for {
         underlyingRef <- repo.createAndOpen(relativePath, notebook, version)